airflow/requirements.txt
apache-airflow-providers-apache-kafka==1.7.0
apache-airflow-providers-apache-spark==5.0.0
apache-airflow-providers-postgres==6.0.0
apache-airflow-providers-slack==9.0.0
apache-airflow-providers-common-sql==1.21.0
confluent-kafka==2.8.0
pandas==2.1.4
pyspark==3.5.0
pyyaml==6.0.2
pymongo==4.11
kafka-python==2.1.5
psycopg2==2.9.10
python-dotenv==1.0.1
slack_sdk==3.34.0
prometheus_client==0.21.1
pendulum==3.0.0
pytest==8.3.4
configparser==7.2.0
pyyaml==6.0.2
user-agents==0.1.0
configparser==7.2.0




